Building_database_apps/requirements.txt
psycopg2-binary==2.9.1
tk==0.1.0
